+++ src/cli/cmd-add.ts
@@ -1,46 +1,46 @@
 import log from "./logger";
 import url from "url";
-import { isPackageUrl, PackageUrl } from "./domain/package-url";
+import { isPackageUrl, PackageUrl } from "../domain/package-url";
 import {
   ManifestLoadError,
   ManifestSaveError,
   tryLoadProjectManifest,
   trySaveProjectManifest,
-} from "./io/project-manifest-io";
-import { EnvParseError, parseEnv } from "./utils/env";
+} from "../io/project-manifest-io";
+import { EnvParseError, parseEnv } from "../utils/env";
 import {
   compareEditorVersion,
   tryParseEditorVersion,
-} from "./domain/editor-version";
-import { makeNpmClient } from "./npm-client";
-import { DomainName } from "./domain/domain-name";
+} from "../domain/editor-version";
+import { makeNpmClient } from "../npm-client";
+import { DomainName } from "../domain/domain-name";
 import {
   makePackageReference,
   PackageReference,
   splitPackageReference,
-} from "./domain/package-reference";
-import { addScope, makeScopedRegistry } from "./domain/scoped-registry";
+} from "../domain/package-reference";
+import { addScope, makeScopedRegistry } from "../domain/scoped-registry";
 import {
   addDependency,
   addTestable,
   mapScopedRegistry,
-} from "./domain/project-manifest";
-import { CmdOptions } from "./types/options";
+} from "../domain/project-manifest";
+import { CmdOptions } from "./options";
 import {
   PackumentResolveError,
   tryResolve,
   VersionNotFoundError,
-} from "./packument-resolving";
-import { SemanticVersion } from "./domain/semantic-version";
-import { fetchPackageDependencies } from "./dependency-resolving";
-import { areArraysEqual } from "./utils/array-utils";
-import { RegistryUrl } from "./domain/registry-url";
-import { PackumentNotFoundError } from "./common-errors";
+} from "../packument-resolving";
+import { SemanticVersion } from "../domain/semantic-version";
+import { fetchPackageDependencies } from "../dependency-resolving";
+import { areArraysEqual } from "../utils/array-utils";
+import { RegistryUrl } from "../domain/registry-url";
+import { PackumentNotFoundError } from "../common-errors";
 import { Err, Ok, Result } from "ts-results-es";
 import { HttpErrorBase } from "npm-registry-fetch";
 import { CustomError } from "ts-custom-error";
-import { logManifestLoadError, logManifestSaveError } from "./error-logging";
-import { targetEditorVersionFor } from "./domain/packument";
+import { logManifestLoadError, logManifestSaveError } from "../error-logging";
+import { targetEditorVersionFor } from "../domain/packument";
 
 export class InvalidPackumentDataError extends CustomError {
   constructor(readonly issue: string) {
