+++ src/cli/cmd-view.ts
@@ -1,18 +1,18 @@
 import chalk from "chalk";
 import log from "./logger";
 import assert from "assert";
-import { tryGetLatestVersion, UnityPackument } from "./domain/packument";
-import { EnvParseError, parseEnv } from "./utils/env";
-import { makeNpmClient } from "./npm-client";
-import { hasVersion, PackageReference } from "./domain/package-reference";
-import { CmdOptions } from "./types/options";
-import { recordKeys } from "./utils/record-utils";
+import { tryGetLatestVersion, UnityPackument } from "../domain/packument";
+import { EnvParseError, parseEnv } from "../utils/env";
+import { makeNpmClient } from "../npm-client";
+import { hasVersion, PackageReference } from "../domain/package-reference";
+import { CmdOptions } from "./options";
+import { recordKeys } from "../utils/record-utils";
 import { Err, Ok, Result } from "ts-results-es";
 
 import {
   PackageWithVersionError,
   PackumentNotFoundError,
-} from "./common-errors";
+} from "../common-errors";
 
 export type ViewOptions = CmdOptions;
