+++ src/cli/logger.ts
@@ -1,5 +1,5 @@
 import npmlog from "npmlog";
-import { tryGetEnv } from "./utils/env-util";
+import { tryGetEnv } from "../utils/env-util";
 
 if (tryGetEnv("NODE_ENV") === "test") {
   npmlog.stream = process.stdout;
